requirements.txt
# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.2
